requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
aiogram==3.4.1
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2
